flask_note/requirements.txt
Flask_Bootstrap
